rename constants skins-animations
