+++ src/constants/skins-animations.js
